GenAI_Search_Engine/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface 
langchain-ollama
numpy==1.26.4
chromadb
langchain_chroma
streamlit
langchain-groq
langchain-core
fastapi
uvicorn
langserve
sse_starlette
httpx
pydantic
transformers
duckdb
duckduckgo-search